requirements.txt
sqlalchemy~=2.0
psycopg2-binary~=2.9
pandas~=2.2
google-auth ~=2.36
google-api-python-client ~=2.152
mysql-connector-python ~=9.3.0
beautifulsoup4 ~=4.13.4
pytest ~=8.3.5
pytest-cov  ~=6.1.1
numpy ~=2.2.6
requests ~=2.32.3